expanded textarea, added sender name, keep draft text
